OL_resolved
Right. That’s it for this task.

A few improvements that I do notice is that the methods could print out
what they are intending to do and automatically have a another method
nested inside to print the pass/fail result.

That’s for interface though but both of them seem functional after
light testing.